pyproject.toml
[tool.poetry]
name = "bitburrow"
version = "0.6.0"
description = "BitBurrow is a set of tools to help you set up and use a VPN server anywhere."
authors = ["bitinerant <bitinerant@users.noreply.github.com>"]
license = "AGPL-3.0-or-later"
readme = "README.md"
repository = "https://github.com/BitBurrow/BitBurrow"
documentation = "https://bitburrow.com/"
packages = [
    {include = "hub"},
    {include = "hub_installer"},
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/BitBurrow/BitBurrow/issues"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.92.0"
slowapi = "==0.1.5"
platformdirs = "^3.0.0"
sqlmodel = "^0.0.8"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
argon2-cffi = "^21.3.0"
PyYAML = "^6.0"
python-socketio = "^5.7.2"

[tool.poetry.dev-dependencies]
black = "^23.1.0"

[tool.poetry.scripts]
bbhub = "hub.hub:entry_point"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
